Fix typing to allow different return values per query
The current typing requires one to define the return type for all queries
when initializing AthenaExpress.

As one wants to execute different queries with different return types they
would have to create new instances of AthenaExpress or ignore the typing
(using `as` or `@ts-ignore`).

By moving the generic parameter from the class to the `query` method, the
expected return type can be set for every call seperatly.

I also added a default parameter `any` so one does not need to define the
return type if they do not with to do so or do not know the exact typing,
while still being able to use the general typing of `QueryResult`.