Extend type support for dates and timestamps
Optionally (defaults to false) cast date, timestamp, and timestamp with
time zone data types to JS Date objects. Query method doesn't expose the
data type forcing developers that want to operate with Date objects to
track which columns are dates and perform the cast themselves.

Dates and timestamps are  casted to JS Date types using Luxon library.
The code uses the method to transform SQL formats, consistent with
Presto/Trino documentation.

The code uses Luxon, and as such, it requires ICU full extension to
work. ICU full mode is available by default starting on Node v13, but
needs to be configured for previous versions of Node.